deploy/requirement.txt
streamlit
torch
torchvision
ultralytics
opencv-python
pillow
matplotlib